Revert "JWT 미들웨어 구현 및 수정"
