+++ build.gradle
@@ -28,11 +28,6 @@ repositories {
     mavenCentral()
 }
 
-tasks.withType(JavaCompile) {
-    options.compilerArgs << "-parameters"
-}
-
-
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-security'
@@ -73,44 +68,9 @@ dependencies {
 
 }
 
-test {
+tasks.named('test') {
     useJUnitPlatform()
-    finalizedBy jacocoTestReport
-
 }
-jacoco {
-    toolVersion = '0.8.7'
-}
-jacocoTestReport {
-    dependsOn test
-
-    reports {
-        xml.required.set(true)
-        html.required.set(true)
-
-        // QueryDSL Q클래스 제외
-        def Qdomains = []
-        for (qPattern in "**/QA".."**/QZ") {
-            Qdomains.add(qPattern + "*")
-        }
-
-        afterEvaluate {
-            classDirectories.setFrom(files(classDirectories.files.collect {
-                fileTree(dir: it,
-                        exclude: [
-                                "**/InitData*",
-                                  "**/global/**"
-                        ,"**/*Application*"] + Qdomains)
-            }))
-        }
-
-        xml.destination file("${buildDir}/jacoco/index.xml")
-        html.destination file("${buildDir}/jacoco/index.html")
-    }
-}
-
-
-
 
 //Querydsl 추가,   동 생성된 Q클래스 gradle clean으로 제거
 clean {
@@ -120,27 +80,5 @@ clean {
 
 
 
-jacocoTestCoverageVerification {
-    violationRules {
-        rule {
-            enabled = true;
-            element = 'CLASS'
 
-            // 라인 커버리지 제한을 90%로 설정
-            limit {
-                counter = 'LINE'
-                value = 'COVEREDRATIO'
-                minimum = 0.80
-            }
 
-            excludes = [
-                    '**.InitData*',
-                    '**.*Application',
-                    '**.domain.**',
-                    '**.global.**'
-
-            ]
-        }
-
-    }
-}
