--- lombok.config
@@ -1,1 +1,0 @@
-lombok.addLombokGeneratedAnnotation = true
