+++ src/main/java/com/study/studyproject/InitData.java
@@ -22,7 +22,6 @@ import java.util.Random;
 
 
 @Component
-@Profile({"local","prod"})
 public class InitData {
 
     @Autowired
