+++ src/main/java/com/study/studyproject/member/controller/AdminController.java
@@ -34,16 +34,17 @@ public class AdminController {
     @GetMapping("/user-all")
     @Operation(summary = "전체 사용자 조회", description = "사용자 검색 조회 및 사용자 전체 조회 가능")
     public Page<UserInfoResponseDto> userAllInfo(
-            @RequestParam(required = false,value = "username") String username,
+            @RequestParam(required = false) String username,
             @PageableDefault(size = 10) Pageable pageable) {
         return adminService.userInfoAll(username,pageable);
     }
 
     @GetMapping("/dash-board")
     @Operation(summary = "관리자 대시보드 ", description = "관리자 정보와 게시글 정보 조회")
-    public AdminDashBoardResponseDto adminDashInfo(@AuthenticationPrincipal UserDetailsImpl userDetails,
+    public AdminDashBoardResponseDto adminDashInfo(@RequestHeader("Access_Token") String token,
                                                    MemberListRequestDto memberListRequestDto, @PageableDefault(size = 10) Pageable pageable) {
-        return adminService.adminDashBoardInfo(userDetails.getMember(),memberListRequestDto,pageable);
+        Long idFromToken = jwtUtil.getIdFromToken(jwtUtil.resolveToken(token));
+        return adminService.adminDashBoardInfo(idFromToken,memberListRequestDto,pageable);
     }
 
     @DeleteMapping("/board/{boardId}")
