+++ src/main/java/com/study/studyproject/member/repository/MemberRepository.java
@@ -14,5 +14,4 @@ import java.util.function.Function;
 public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
 
     Optional<Member> findByEmail(String email);
-    Optional<Member> findByNickname(String email);
 }
