+++ src/main/java/com/study/studyproject/member/service/AdminServiceImpl.java
@@ -36,7 +36,8 @@ public class AdminServiceImpl implements AdminService{
 
     }
 
-    public AdminDashBoardResponseDto adminDashBoardInfo(Member member, MemberListRequestDto memberListRequestDto, Pageable pageable) {
+    public AdminDashBoardResponseDto adminDashBoardInfo(Long id, MemberListRequestDto memberListRequestDto, Pageable pageable) {
+        Member member = memberRepository.findById(id).orElseThrow(() -> new UserNotFoundException("사용자를 찾지 못했습니다."));
         Page<ListResponseDto> listResponseDtos = myPageQueryRepository.MyPageListPage(memberListRequestDto, pageable, null,getAdmin);
         return AdminDashBoardResponseDto.of(member, listResponseDtos);
     }
