+++ src/main/java/com/study/studyproject/reply/service/ReplyService.java
@@ -9,7 +9,7 @@ import jakarta.servlet.http.HttpServletRequest;
 public interface ReplyService {
 
 
-    void insert(Member member, ReplyRequestDto replyRequestDto);
+    void insert(Long idFromToken, ReplyRequestDto replyRequestDto);
 
     //수정
     void updateReply(UpdateReplyRequest updateReplyRequest);
