+++ src/test/java/com/study/studyproject/reply/service/ReplyServiceImplTest.java
@@ -63,7 +63,7 @@ class ReplyServiceImplTest {
         ReplyRequestDto replyOne = new ReplyRequestDto(boardCreate.getId(), null, "댓글 내용");
 
         //when
-        replyService.insert(member1, replyOne);
+        replyService.insert(member1.getId(), replyOne);
 
         //then
         List<Reply> all = replyRepository.findAll();
@@ -88,7 +88,7 @@ class ReplyServiceImplTest {
         ReplyRequestDto replyOne = new ReplyRequestDto(board.getId(), parentReply.getId(), "댓글 내용");
 
         //when
-        replyService.insert(member1, replyOne);
+        replyService.insert(member1.getId(), replyOne);
 
         //then
         List<Reply> all = replyRepository.findAll();
